:recycle: Refactor browse screen header
Browse screenのheaderをcomponent化
検索欄およびiconをdark modeに対応
