+++ .idea/misc.xml
@@ -25,6 +25,9 @@
         <entry key="../../../../layout/compose-model-1644052312678.xml" value="0.5" />
         <entry key="../../../../layout/compose-model-1644052321111.xml" value="2.0" />
         <entry key="../../../../layout/compose-model-1644052484073.xml" value="1.0" />
+        <entry key="../../../../layout/compose-model-1644060295460.xml" value="0.1540551839464883" />
+        <entry key="../../../../layout/compose-model-1644060513202.xml" value="2.0" />
+        <entry key="../../../../layout/compose-model-1644065117132.xml" value="2.0" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.18541666666666667" />
         <entry key="app/src/main/res/drawable/baseline_public.xml" value="0.21354166666666666" />
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.1" />
